+++ e2e/_suites/fleet/ingest_manager_test.go
@@ -89,37 +89,23 @@ func setUpSuite() {
 			kibanaClient: kibanaClient,
 			Installers:   map[string]installer.ElasticAgentInstaller{}, // do not pre-initialise the map
 		},
-		//StandAlone: &StandAloneTestSuite{
-		//	kibanaClient: kibanaClient,
-		//},
 	}
 }
 
 func InitializeIngestManagerTestScenario(ctx *godog.ScenarioContext) {
 	ctx.BeforeScenario(func(*messages.Pickle) {
 		log.Trace("Before Fleet scenario")
-
-		//imts.StandAlone.Cleanup = false
-
 		imts.Fleet.beforeScenario()
 	})
 
 	ctx.AfterScenario(func(*messages.Pickle, error) {
 		log.Trace("After Fleet scenario")
-
-		//		if imts.StandAlone.Cleanup {
-		//			imts.StandAlone.afterScenario()
-		//		}
-
-		//if imts.Fleet.Cleanup {
 		imts.Fleet.afterScenario()
-		//}
 	})
 
 	ctx.Step(`^the "([^"]*)" process is in the "([^"]*)" state on the host$`, imts.processStateOnTheHost)
 
 	imts.Fleet.contributeSteps(ctx)
-	//	imts.StandAlone.contributeSteps(ctx)
 }
 
 func InitializeIngestManagerTestSuite(ctx *godog.TestSuiteContext) {
