+++ e2e/_suites/fleet/stand-alone.go
@@ -7,128 +7,94 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/cenkalti/backoff/v4"
 	"github.com/elastic/e2e-testing/cli/config"
+	"github.com/elastic/e2e-testing/internal/common"
+	"github.com/elastic/e2e-testing/internal/compose"
+	"github.com/elastic/e2e-testing/internal/docker"
+	"github.com/elastic/e2e-testing/internal/installer"
+	"github.com/elastic/e2e-testing/internal/shell"
+	"github.com/elastic/e2e-testing/internal/utils"
 	"path"
 	"strings"
 	"time"
 
-	"github.com/cenkalti/backoff/v4"
-	"github.com/elastic/e2e-testing/internal/common"
-	"github.com/elastic/e2e-testing/internal/compose"
-	"github.com/elastic/e2e-testing/internal/docker"
 	"github.com/elastic/e2e-testing/internal/elasticsearch"
-	"github.com/elastic/e2e-testing/internal/installer"
-	"github.com/elastic/e2e-testing/internal/kibana"
-	"github.com/elastic/e2e-testing/internal/shell"
-	"github.com/elastic/e2e-testing/internal/utils"
-	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 )
 
-// StandAloneTestSuite represents the scenarios for Stand-alone-mode
-type StandAloneTestSuite struct {
-	Cleanup     bool
-	Hostname    string
-	Image       string
-	FleetPolicy kibana.Policy
-	// date controls for queries
-	AgentStoppedDate             time.Time
-	RuntimeDependenciesStartDate time.Time
-	kibanaClient                 *kibana.Client
+func (fts *FleetTestSuite) aStandaloneAgentIsDeployed(image string) error {
+	fts.StandAlone = true
+	return fts.startAgent(image, "", nil)
 }
 
-// afterScenario destroys the state created by a scenario
-func (sats *StandAloneTestSuite) afterScenario() {
-	serviceManager := compose.NewServiceManager()
-	serviceName := common.ElasticAgentServiceName
-
-	if log.IsLevelEnabled(log.DebugLevel) {
-		_ = sats.getContainerLogs()
-	}
-
-	developerMode := shell.GetEnvBool("DEVELOPER_MODE")
-	if !developerMode {
-		_ = serviceManager.RemoveServicesFromCompose(context.Background(), common.FleetProfileName, []string{serviceName}, common.ProfileEnv)
-	} else {
-		log.WithField("service", serviceName).Info("Because we are running in development mode, the service won't be stopped")
-	}
-
-	sats.kibanaClient.DeleteAllPolicies(sats.FleetPolicy)
-}
-
-/*
-func (sats *StandAloneTestSuite) contributeSteps(s *godog.ScenarioContext) {
-	s.Step(`^a "([^"]*)" stand-alone agent is deployed$`, sats.aStandaloneAgentIsDeployed)
-	s.Step(`^a "([^"]*)" stand-alone agent is deployed with fleet server mode$`, sats.bootstrapFleetServerFromAStandaloneAgent)
-	s.Step(`^a "([^"]*)" stand-alone agent is deployed with fleet server mode on cloud$`, sats.aStandaloneAgentIsDeployedWithFleetServerModeOnCloud)
-	s.Step(`^there is new data in the index from agent$`, sats.thereIsNewDataInTheIndexFromAgent)
-	s.Step(`^the "([^"]*)" docker container is stopped$`, sats.theDockerContainerIsStopped)
-	s.Step(`^there is no new data in the index after agent shuts down$`, sats.thereIsNoNewDataInTheIndexAfterAgentShutsDown)
-	s.Step(`^the stand-alone agent is listed in Fleet as "([^"]*)"$`, sats.theStandaloneAgentIsListedInFleetWithStatus)
-	s.Step(`^the "([^"]*)" integration is added to the policy$`, sats.theIntegrationIsAddedToThePolicy)
-	s.Step(`^the "([^"]*)" datasource is shown in the policy$`, sats.thePolicyShowsTheDatasourceAdded)
-}
-*/
-
-func (sats *StandAloneTestSuite) theIntegrationIsAddedToThePolicy(packageName string) error {
-	return theIntegrationIsOperatedInThePolicy(sats.kibanaClient, sats.FleetPolicy, packageName, "added")
-}
-
-func (sats *StandAloneTestSuite) thePolicyShowsTheDatasourceAdded(packageName string) error {
-	return nil
-	//return thePolicyShowsTheDatasourceAdded(sats.kibanaClient, sats.FleetPolicy, packageName)
-}
-
-func (sats *StandAloneTestSuite) aStandaloneAgentIsDeployedWithFleetServerModeOnCloud(image string) error {
-	fleetPolicy, err := sats.kibanaClient.GetDefaultPolicy(true)
+func (fts *FleetTestSuite) bootstrapFleetServerFromAStandaloneAgent(image string) error {
+	fts.StandAlone = true
+	fleetPolicy, err := fts.kibanaClient.GetDefaultPolicy(true)
 	if err != nil {
 		return err
 	}
-	sats.FleetPolicy = fleetPolicy
-	volume := path.Join(config.OpDir(), "compose", "services", "elastic-agent", "apm-legacy")
-	return sats.startAgent(image, "docker-compose-cloud.yml", map[string]string{"apmVolume": volume})
+	fts.FleetServerPolicy = fleetPolicy
+	return fts.startAgent(image, "", map[string]string{"fleetServerMode": "1"})
 }
 
-func (sats *StandAloneTestSuite) theStandaloneAgentIsListedInFleetWithStatus(desiredStatus string) error {
-	waitForAgents := func() error {
-		agents, err := sats.kibanaClient.ListAgents()
-		if err != nil {
+func (fts *FleetTestSuite) aStandaloneAgentIsDeployedWithFleetServerModeOnCloud(image string) error {
+	fts.StandAlone = true
+	fleetPolicy, err := fts.kibanaClient.GetDefaultPolicy(true)
+	if err != nil {
+		return err
+	}
+	fts.FleetServerPolicy = fleetPolicy
+	volume := path.Join(config.OpDir(), "compose", "services", "elastic-agent", "apm-legacy")
+	return fts.startAgent(image, "docker-compose-cloud.yml", map[string]string{"apmVolume": volume})
+}
+
+func (fts *FleetTestSuite) thereIsNewDataInTheIndexFromAgent() error {
+	maxTimeout := time.Duration(common.TimeoutFactor) * time.Minute * 2
+	minimumHitsCount := 50
+
+	result, err := searchAgentData(fts.Hostname, fts.RuntimeDependenciesStartDate, minimumHitsCount, maxTimeout)
+	if err != nil {
+		return err
+	}
+
+	log.Tracef("Search result: %v", result)
+
+	return elasticsearch.AssertHitsArePresent(result)
+}
+
+func (fts *FleetTestSuite) theDockerContainerIsStopped(serviceName string) error {
+	serviceManager := compose.NewServiceManager()
+
+	err := serviceManager.RemoveServicesFromCompose(context.Background(), common.FleetProfileName, []string{serviceName}, common.ProfileEnv)
+	if err != nil {
+		return err
+	}
+	fts.AgentStoppedDate = time.Now().UTC()
+
+	return nil
+}
+
+func (fts *FleetTestSuite) thereIsNoNewDataInTheIndexAfterAgentShutsDown() error {
+	maxTimeout := time.Duration(30) * time.Second
+	minimumHitsCount := 1
+
+	result, err := searchAgentData(fts.Hostname, fts.AgentStoppedDate, minimumHitsCount, maxTimeout)
+	if err != nil {
+		if strings.Contains(err.Error(), "type:index_not_found_exception") {
 			return err
 		}
 
-		if len(agents) == 0 {
-			return errors.New("No agents found")
-		}
-
-		agentZero := agents[0]
-		hostname := agentZero.LocalMetadata.Host.HostName
-
-		return theAgentIsListedInFleetWithStatus(desiredStatus, hostname)
+		log.WithFields(log.Fields{
+			"error": err,
+		}).Info("No documents were found for the Agent in the index after it stopped")
+		return nil
 	}
-	maxTimeout := time.Duration(common.TimeoutFactor) * time.Minute * 2
-	exp := common.GetExponentialBackOff(maxTimeout)
 
-	err := backoff.Retry(waitForAgents, exp)
-	if err != nil {
-		return err
-	}
-	return nil
+	return elasticsearch.AssertHitsAreNotPresent(result)
 }
 
-func (sats *StandAloneTestSuite) bootstrapFleetServerFromAStandaloneAgent(image string) error {
-	fleetPolicy, err := sats.kibanaClient.GetDefaultPolicy(true)
-	if err != nil {
-		return err
-	}
-	sats.FleetPolicy = fleetPolicy
-	return sats.startAgent(image, "", map[string]string{"fleetServerMode": "1"})
-}
-
-func (sats *StandAloneTestSuite) aStandaloneAgentIsDeployed(image string) error {
-	return sats.startAgent(image, "", nil)
-}
-
-func (sats *StandAloneTestSuite) startAgent(image string, composeFilename string, env map[string]string) error {
+func (fts *FleetTestSuite) startAgent(image string, composeFilename string, env map[string]string) error {
 
 	log.Trace("Deploying an agent to Fleet")
 
@@ -179,11 +145,10 @@ func (sats *StandAloneTestSuite) startAgent(image string, composeFilename string
 		return err
 	}
 
-	sats.Image = image
-	sats.Hostname = hostname
-	sats.Cleanup = true
+	fts.Image = image
+	fts.Hostname = hostname
 
-	err = sats.installTestTools(containerName)
+	err = fts.installTestTools(containerName)
 	if err != nil {
 		return err
 	}
@@ -191,23 +156,41 @@ func (sats *StandAloneTestSuite) startAgent(image string, composeFilename string
 	return nil
 }
 
-func (sats *StandAloneTestSuite) getContainerLogs() error {
-	serviceManager := compose.NewServiceManager()
+func (fts *FleetTestSuite) theIntegrationIsAddedToThePolicy(packageName string) error {
+	return theIntegrationIsOperatedInThePolicy(fts.kibanaClient, fts.FleetServerPolicy, packageName, "added")
+}
 
-	profile := common.FleetProfileName
-	serviceName := common.ElasticAgentServiceName
+func (fts *FleetTestSuite) thePolicyShowsTheDatasourceAdded(packageName string) error {
+	fts.StandAlone = true
+	log.WithFields(log.Fields{
+		"policyID": fts.FleetServerPolicy.ID,
+		"package":  packageName,
+	}).Trace("Checking if the policy shows the package added")
 
-	composes := []string{
-		profile,     // profile name
-		serviceName, // agent service
+	maxTimeout := time.Minute
+	retryCount := 1
+
+	exp := common.GetExponentialBackOff(maxTimeout)
+
+	configurationIsPresentFn := func() error {
+		packagePolicy, err := fts.kibanaClient.GetIntegrationFromAgentPolicy(packageName, fts.FleetServerPolicy)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"packagePolicy": packagePolicy,
+				"policy":        fts.FleetServerPolicy,
+				"retry":         retryCount,
+				"error":         err,
+			}).Warn("The integration was not found in the policy")
+			retryCount++
+			return err
+		}
+
+		retryCount++
+		return err
 	}
-	err := serviceManager.RunCommand(profile, composes, []string{"logs", serviceName}, common.ProfileEnv)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"error":   err,
-			"service": serviceName,
-		}).Error("Could not retrieve Elastic Agent logs")
 
+	err := backoff.Retry(configurationIsPresentFn, exp)
+	if err != nil {
 		return err
 	}
 
@@ -217,8 +200,8 @@ func (sats *StandAloneTestSuite) getContainerLogs() error {
 // installTestTools we need the container name because we use the Docker Client instead of Docker Compose
 // we are going to install those tools we use in the test framework for checking
 // and verifications
-func (sats *StandAloneTestSuite) installTestTools(containerName string) error {
-	if sats.Image != "ubi8" {
+func (fts *FleetTestSuite) installTestTools(containerName string) error {
+	if fts.Image != "ubi8" {
 		return nil
 	}
 
@@ -245,51 +228,6 @@ func (sats *StandAloneTestSuite) installTestTools(containerName string) error {
 	}).Debug("Test tools installed")
 
 	return nil
-}
-
-func (sats *StandAloneTestSuite) thereIsNewDataInTheIndexFromAgent() error {
-	maxTimeout := time.Duration(common.TimeoutFactor) * time.Minute * 2
-	minimumHitsCount := 50
-
-	result, err := searchAgentData(sats.Hostname, sats.RuntimeDependenciesStartDate, minimumHitsCount, maxTimeout)
-	if err != nil {
-		return err
-	}
-
-	log.Tracef("Search result: %v", result)
-
-	return elasticsearch.AssertHitsArePresent(result)
-}
-
-func (sats *StandAloneTestSuite) theDockerContainerIsStopped(serviceName string) error {
-	serviceManager := compose.NewServiceManager()
-
-	err := serviceManager.RemoveServicesFromCompose(context.Background(), common.FleetProfileName, []string{serviceName}, common.ProfileEnv)
-	if err != nil {
-		return err
-	}
-	sats.AgentStoppedDate = time.Now().UTC()
-
-	return nil
-}
-
-func (sats *StandAloneTestSuite) thereIsNoNewDataInTheIndexAfterAgentShutsDown() error {
-	maxTimeout := time.Duration(30) * time.Second
-	minimumHitsCount := 1
-
-	result, err := searchAgentData(sats.Hostname, sats.AgentStoppedDate, minimumHitsCount, maxTimeout)
-	if err != nil {
-		if strings.Contains(err.Error(), "type:index_not_found_exception") {
-			return err
-		}
-
-		log.WithFields(log.Fields{
-			"error": err,
-		}).Info("No documents were found for the Agent in the index after it stopped")
-		return nil
-	}
-
-	return elasticsearch.AssertHitsAreNotPresent(result)
 }
 
 func searchAgentData(hostname string, startDate time.Time, minimumHitsCount int, maxTimeout time.Duration) (elasticsearch.SearchResult, error) {
